sync workflow files (#50)
* Add GitHub Actions workflow for .NET build and release (#47)

* Add GitHub Actions workflow for .NET build and release

This automatically builds and packages an exe into Releases

* Update release workflow to use solution and project files

earlier one didnt work because csproj and sln was in same folder. Needed to specify which one to use. Used sln for restore and build. Csproj for publish to just get an exe

* Add GitHub Actions workflow for .NET project (#48)

builds the project on pull requests in dev. Makes sure the project can build.